[X]update_03_02_2026/requirements.txt
tensorflow
numpy
matplotlib
pandas
pyyaml
tf2onnx
scipy